S3: Enable bucket logging
